Revert "Fix completion candidates not updating."
I misunderstood how deoplete works in regards to caching/filtering and the
change to how sources that have not set is_volatile are handled is indeed
desireable.

If is_volatile is not set for the source then gather_candidates is only called
for the initial completion (i.e. when min_pattern_length is reached) and further
modifications of the PUM candidates are managed by filters.

This is how fuzzy filters are able to work.
If is_volatile is set then the candidates from 'complete --do-complete' are updated for
all user input, which means we will only ever have exact prefix matches (e.g. no
fuzzy filtering).

This reverts commit a8369cb9d790c090ab38a4c91c38b9754debec8a.